Refactor MarkdownDocs out of CsRepo and RubyRepo
This can be reused across both
