+++ lib/teacher_services_tech_docs.rb
@@ -23,6 +23,7 @@ require "teacher_services_tech_docs/pages_by_category"
 
 # business
 require "teacher_services_tech_docs/github/client"
+require "teacher_services_tech_docs/github/markdown_docs"
 require "teacher_services_tech_docs/github/ruby_repo"
 require "teacher_services_tech_docs/github/cs_repo"
 require "teacher_services_tech_docs/github/file"
